// apps/web/src/generated/api/models/commentAuthorItemResponseDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * ThreadNest API
 * OpenAPI spec version: 1.0.0
 */
import type { CommentAttachmentResponseDto } from './commentAttachmentResponseDto'
import type { CommentAuthorThreadReferenceDto } from './commentAuthorThreadReferenceDto'
import type { NestReferencetDto } from './nestReferencetDto'

export interface CommentAuthorItemResponseDto {
  /** Comment ID */
  id: string
  /** Comment content */
  content: string
  /** Creation timestamp */
  createdAt: string
  /** The thread this comment was posted in */
  thread: CommentAuthorThreadReferenceDto
  /** The nest this comment belongs to */
  nest: NestReferencetDto
  /**
     * The comment's attachment, if it has one
     * @nullable
     */
  attachment: CommentAttachmentResponseDto | null
}